LayoutTests/js/ShadowRealm-iframe-detach-expected.txt

Test to ensure correct behaviour of Shadow Realm objects that escape their originating iframe

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


[object Window]
[object Window]
detaching iframe
[object ShadowRealm]
dropping realm value
PASS () => checkFn2(wrappedLog("shadowRealm")) is true
PASS successfullyParsed is true

TEST COMPLETE
